Rename to align with usage
