+++ src/app/flows/fines/fines-mac/services/fines-mac-payload/constants/fines-mac-defendant-debtor-details-alias-payload.constant.ts
@@ -1,4 +1,4 @@
-import { IFinesMacPayloadAccountDefendantDebtorDetailAliasComplete } from '../utils/interfaces/fines-mac-payload-defendant-debtor-detail-alias-complete.interface';
+import { IFinesMacPayloadAccountDefendantDebtorDetailAliasComplete } from '../utils/interfaces/fines-mac-payload-account-defendant-debtor-detail-alias-complete.interface';
 
 export const FINES_MAC_DEFENDANT_DEBTOR_DETAILS_ALIAS_PAYLOAD: IFinesMacPayloadAccountDefendantDebtorDetailAliasComplete =
   {
