+++ src/app/flows/fines/fines-mac/services/fines-mac-payload/constants/fines-mac-defendant-parent-guardian-payload.constant.ts
@@ -1,4 +1,4 @@
-import { IFinesMacPayloadAccountDefendantParentGuardianComplete } from '../utils/interfaces/fines-mac-payload-defendant-parent-guardian-complete.interface';
+import { IFinesMacPayloadAccountDefendantParentGuardianComplete } from '../utils/interfaces/fines-mac-payload-account-defendant-parent-guardian-complete.interface';
 import { FINES_MAC_DEFENDANT_DEBTOR_DETAILS_PAYLOAD } from './fines-mac-defendant-debtor-details-payload.constant';
 
 export const FINES_MAC_DEFENDANT_PARENT_GUARDIAN_PAYLOAD: IFinesMacPayloadAccountDefendantParentGuardianComplete = {
