+++ src/app/flows/fines/fines-mac/services/fines-mac-payload/constants/fines-mac-defendant-payload.constant.ts
@@ -1,4 +1,4 @@
-import { IFinesMacPayloadAccountDefendantComplete } from '../utils/interfaces/fines-mac-payload-defendant-complete.interface';
+import { IFinesMacPayloadAccountDefendantComplete } from '../utils/interfaces/fines-mac-payload-account-defendant-complete.interface';
 import { FINES_MAC_DEFENDANT_DEBTOR_DETAILS_PAYLOAD } from './fines-mac-defendant-debtor-details-payload.constant';
 import { FINES_MAC_DEFENDANT_PARENT_GUARDIAN_PAYLOAD } from './fines-mac-defendant-parent-guardian-payload.constant';
