+++ src/app/flows/fines/fines-mac/services/fines-mac-payload/interfaces/fines-mac-payload-account.interface.ts
@@ -1,12 +1,12 @@
-import { IFinesMacPayloadAccountAccountNote } from '../utils/interfaces/fines-mac-payload-account-note.interface';
-import { IFinesMacPayloadAccountDefendantComplete } from '../utils/interfaces/fines-mac-payload-defendant-complete.interface';
+import { IFinesMacPayloadAccountAccountNote } from '../utils/interfaces/fines-mac-payload-account-account-note.interface';
+import { IFinesMacPayloadAccountDefendantComplete } from '../utils/interfaces/fines-mac-payload-account-defendant-complete.interface';
 import { IFinesMacPayloadAccountAccountInitial } from './fines-mac-payload-account-initial.interface';
-import { IFinesMacPaymentTermsPayload } from '../utils/interfaces/fines-mac-payment-terms-payload.interface';
+import { IFinesMacPayloadAccountPaymentTerms } from '../utils/interfaces/fines-mac-payload-account-payment-terms.interface';
 
 export interface IFinesMacPayloadAccount extends IFinesMacPayloadAccountAccountInitial {
   defendant: IFinesMacPayloadAccountDefendantComplete;
   offences: null;
   fp_ticket_detail: null;
-  payment_terms: IFinesMacPaymentTermsPayload;
+  payment_terms: IFinesMacPayloadAccountPaymentTerms;
   account_notes: IFinesMacPayloadAccountAccountNote[] | null;
 }
