+++ src/app/flows/fines/fines-mac/services/fines-mac-payload/utils/fines-mac-payload-account-notes.utils.spec.ts
@@ -1,6 +1,6 @@
 import { buildAccountNotesPayload } from './fines-mac-payload-account-notes.utils';
 import { IFinesMacAccountCommentsNotesState } from '../../../fines-mac-account-comments-notes/interfaces/fines-mac-account-comments-notes-state.interface';
-import { IFinesMacPayloadAccountAccountNote } from './interfaces/fines-mac-payload-account-note.interface';
+import { IFinesMacPayloadAccountAccountNote } from './interfaces/fines-mac-payload-account-account-note.interface';
 
 describe('buildAccountNotesPayload', () => {
   it('should return an array of account notes when comments and notes are provided', () => {
