+++ src/app/flows/fines/fines-mac/services/fines-mac-payload/utils/fines-mac-payload-account-notes.utils.ts
@@ -1,5 +1,5 @@
 import { IFinesMacAccountCommentsNotesState } from '../../../fines-mac-account-comments-notes/interfaces/fines-mac-account-comments-notes-state.interface';
-import { IFinesMacPayloadAccountAccountNote } from './interfaces/fines-mac-payload-account-note.interface';
+import { IFinesMacPayloadAccountAccountNote } from './interfaces/fines-mac-payload-account-account-note.interface';
 
 /**
  * Builds an account note object for fines management.
