+++ src/app/flows/fines/fines-mac/services/fines-mac-payload/utils/interfaces/fines-mac-payload-account-defendant-company-debtor-details.interface.ts
@@ -1,4 +1,4 @@
-import { IFinesMacPayloadAccountDefendantCompanyDebtorDetailsAlias } from './fines-mac-payload-defendant-company-debtor-details-alias.interface';
+import { IFinesMacPayloadAccountDefendantCompanyDebtorDetailsAlias } from './fines-mac-payload-account-defendant-company-debtor-details-alias.interface';
 
 export interface IFinesMacPayloadAccountDefendantCompanyDebtorDetails {
   document_language: string | null;
