+++ src/app/flows/fines/fines-mac/services/fines-mac-payload/utils/interfaces/fines-mac-payload-account-defendant-company.interface.ts
@@ -1,4 +1,4 @@
-import { IFinesMacPayloadAccountDefendantCompanyDebtorDetails } from './fines-mac-payload-defendant-company-debtor-details.interface';
+import { IFinesMacPayloadAccountDefendantCompanyDebtorDetails } from './fines-mac-payload-account-defendant-company-debtor-details.interface';
 
 export interface IFinesMacPayloadAccountDefendantCompany {
   company_flag: boolean | null;
