+++ src/app/flows/fines/fines-mac/services/fines-mac-payload/utils/interfaces/fines-mac-payload-account-defendant-debtor-detail-complete.interface.ts
@@ -1,6 +1,6 @@
-import { IFinesMacPayloadAccountDefendantIndividualDebtorDetailsAlias } from './fines-mac-payload-defendant-individual-debtor-details-alias.interface';
+import { IFinesMacPayloadAccountDefendantDebtorDetailAliasComplete } from './fines-mac-payload-account-defendant-debtor-detail-alias-complete.interface';
 
-export interface IFinesMacPayloadAccountDefendantIndividualDebtorDetails {
+export interface IFinesMacPayloadAccountDefendantDebtorDetailComplete {
   vehicle_make: string | null;
   vehicle_registration_mark: string | null;
   document_language: string | null;
@@ -15,5 +15,5 @@ export interface IFinesMacPayloadAccountDefendantIndividualDebtorDetails {
   employer_post_code: string | null;
   employer_telephone_number: string | null;
   employer_email_address: string | null;
-  aliases: IFinesMacPayloadAccountDefendantIndividualDebtorDetailsAlias[] | null;
+  aliases: IFinesMacPayloadAccountDefendantDebtorDetailAliasComplete[] | null;
 }
