+++ src/app/flows/fines/fines-mac/services/fines-mac-payload/utils/interfaces/fines-mac-payload-account-defendant-individual-debtor-details-alias.interface.ts
