+++ src/app/flows/fines/fines-mac/services/fines-mac-payload/utils/interfaces/fines-mac-payload-account-payment-terms-enforcement-result-response.interface.ts
@@ -1,0 +1,4 @@
+export interface IFinesMacPayloadAccountPaymentTermsEnforcementResultResponse {
+  parameter_name: string | null;
+  response: string | null;
+}
