+++ src/app/flows/fines/fines-mac/services/fines-mac-payload/utils/interfaces/fines-mac-payload-account-payment-terms-enforcement.interface.ts
@@ -1,0 +1,6 @@
+import { IFinesMacPayloadAccountPaymentTermsEnforcementResultResponse } from './fines-mac-payload-account-payment-terms-enforcement-result-response.interface';
+
+export interface IFinesMacPayloadAccountPaymentTermsEnforcement {
+  result_id: string | null;
+  enforcement_result_responses: IFinesMacPayloadAccountPaymentTermsEnforcementResultResponse[] | null;
+}
