+++ src/app/flows/fines/fines-mac/services/fines-mac-payload/utils/interfaces/fines-mac-payload-account-payment-terms.interface.ts
@@ -1,0 +1,11 @@
+import { IFinesMacPayloadAccountPaymentTermsEnforcement } from './fines-mac-payload-account-payment-terms-enforcement.interface';
+
+export interface IFinesMacPayloadAccountPaymentTerms {
+  payment_terms_type_code: string | null;
+  effective_date: string | null;
+  instalment_period: string | null;
+  lump_sum_amount: number | null;
+  instalment_amount: number | null;
+  default_days_in_jail: number | null;
+  enforcements: IFinesMacPayloadAccountPaymentTermsEnforcement[] | null;
+}
