--- src/app/flows/fines/fines-mac/services/fines-mac-payload/utils/interfaces/fines-mac-payment-terms-enforcement-payload.interface.ts
@@ -1,6 +1,0 @@
-import { IFinesMacPaymentTermsEnforcementResultResponsePayload } from './fines-mac-payment-terms-enforcement-result-response-payload.interface';
-
-export interface IFinesMacPaymentTermsEnforcementPayload {
-  result_id: string | null;
-  enforcement_result_responses: IFinesMacPaymentTermsEnforcementResultResponsePayload[] | null;
-}
