--- src/app/flows/fines/fines-mac/services/fines-mac-payload/utils/interfaces/fines-mac-payment-terms-enforcement-result-response-payload.interface.ts
@@ -1,4 +1,0 @@
-export interface IFinesMacPaymentTermsEnforcementResultResponsePayload {
-  parameter_name: string | null;
-  response: string | null;
-}
