+++ src/app/flows/fines/fines-mac/services/fines-mac-payload/utils/mocks/fines-mac-payload-account-defendant-company-with-aliases.mock.ts
@@ -1,4 +1,4 @@
-import { IFinesMacPayloadAccountDefendantCompany } from '../interfaces/fines-mac-payload-defendant-company.interface';
+import { IFinesMacPayloadAccountDefendantCompany } from '../interfaces/fines-mac-payload-account-defendant-company.interface';
 import { FINES_MAC_PAYLOAD_ACCOUNT_DEFENDANT_COMPANY_MOCK } from './fines-mac-payload-account-defendant-company.mock';
 
 export const FINES_MAC_PAYLOAD_ACCOUNT_DEFENDANT_COMPANY_WITH_ALIASES_MOCK: IFinesMacPayloadAccountDefendantCompany = {
