+++ src/app/flows/fines/fines-mac/services/fines-mac-payload/utils/mocks/fines-mac-payload-account-defendant-company.mock.ts
@@ -1,4 +1,4 @@
-import { IFinesMacPayloadAccountDefendantCompany } from '../interfaces/fines-mac-payload-defendant-company.interface';
+import { IFinesMacPayloadAccountDefendantCompany } from '../interfaces/fines-mac-payload-account-defendant-company.interface';
 
 export const FINES_MAC_PAYLOAD_ACCOUNT_DEFENDANT_COMPANY_MOCK: IFinesMacPayloadAccountDefendantCompany = {
   company_flag: true,
