+++ src/app/flows/fines/fines-mac/services/fines-mac-payload/utils/mocks/fines-mac-payload-account-defendant-parent-guardian-complete.mock.ts
@@ -1,7 +1,7 @@
 import { FINES_MAC_DEFENDANT_DEBTOR_DETAILS_PAYLOAD } from '../../constants/fines-mac-defendant-debtor-details-payload.constant';
 import { FINES_MAC_DEFENDANT_PARENT_GUARDIAN_PAYLOAD } from '../../constants/fines-mac-defendant-parent-guardian-payload.constant';
 import { FINES_MAC_DEFENDANT_PAYLOAD } from '../../constants/fines-mac-defendant-payload.constant';
-import { IFinesMacPayloadAccountDefendantComplete } from '../interfaces/fines-mac-payload-defendant-complete.interface';
+import { IFinesMacPayloadAccountDefendantComplete } from '../interfaces/fines-mac-payload-account-defendant-complete.interface';
 import { FINES_MAC_PAYLOAD_ACCOUNT_DEFENDANT_PARENT_GUARDIAN_MOCK } from './fines-mac-payload-account-defendant-parent-guardian.mock';
 
 export const FINES_MAC_PAYLOAD_ACCOUNT_DEFENDANT_PARENT_GUARDIAN_COMPLETE_MOCK: IFinesMacPayloadAccountDefendantComplete =
