+++ src/app/flows/fines/fines-mac/services/fines-mac-payload/utils/mocks/fines-mac-payload-account-defendant-parent-guardian-with-alias.mock.ts
@@ -1,4 +1,4 @@
-import { IFinesMacPayloadAccountDefendantParentGuardian } from '../interfaces/fines-mac-payload-defendant-parent-guardian.interface';
+import { IFinesMacPayloadAccountDefendantParentGuardian } from '../interfaces/fines-mac-payload-account-defendant-parent-guardian.interface';
 import { FINES_MAC_PAYLOAD_ACCOUNT_DEFENDANT_PARENT_GUARDIAN_MOCK } from './fines-mac-payload-account-defendant-parent-guardian.mock';
 
 export const FINES_MAC_PAYLOAD_ACCOUNT_DEFENDANT_PARENT_GUARDIAN_WITH_ALIAS_MOCK: IFinesMacPayloadAccountDefendantParentGuardian =
